Rename package from `wimae` to `contrawimae` and update all related references in documentation, examples, and tests.
